These are completed: * Research how to have a delete button in the list by the specific item -> Paul * Add delete functionality to the default list -> Paul
I had to add a custom adapter for the ListView, and I used the native list instead of converting it to an ArrayList<String>.  This allowed for the delete button to act on the actual list.  Then just needed to make sure and save the list, so it wrote it to the JSON file.